junk

+++ src/USMapGenerator/Generator.cpp
@@ -423,7 +423,7 @@ namespace RC::OutTheShade
         std::streampos extEndPos = Buffer.GetBuffer().tellp();
 
         Buffer.GetBuffer().seekp(extStartPos);
-        Buffer.GetBuffer().seekp(-sizeof(uint32_t), std::ios_base::cur);
+        Buffer.GetBuffer().seekp(-static_cast<int32>(sizeof(uint32)), std::ios_base::cur);
         Buffer.Write<uint32_t>(extEndPos - extStartPos);
         Buffer.GetBuffer().seekp(extEndPos);
 
@@ -465,7 +465,7 @@ namespace RC::OutTheShade
         extEndPos = Buffer.GetBuffer().tellp();
 
         Buffer.GetBuffer().seekp(extStartPos);
-        Buffer.GetBuffer().seekp(-sizeof(uint32_t), std::ios_base::cur);
+        Buffer.GetBuffer().seekp(-static_cast<int32_t>(sizeof(uint32_t)), std::ios_base::cur);
         Buffer.Write<uint32_t>(extEndPos - extStartPos);
         Buffer.GetBuffer().seekp(extEndPos);
 
@@ -491,7 +491,7 @@ namespace RC::OutTheShade
         extEndPos = Buffer.GetBuffer().tellp();
 
         Buffer.GetBuffer().seekp(extStartPos);
-        Buffer.GetBuffer().seekp(-sizeof(uint32_t), std::ios_base::cur);
+        Buffer.GetBuffer().seekp(-static_cast<int32_t>(sizeof(uint32_t)), std::ios_base::cur);
         Buffer.Write<uint32_t>(extEndPos - extStartPos);
         Buffer.GetBuffer().seekp(extEndPos);
